Make `Arguments` pickle-able
What's preventing pickling is the `Switch` type, stored in
`__annotations__`. `Switch` references the dummy type, previously
declared under the name `--invalid-type--`. The intent was to make sure
that no class with such name exists, however it also prevented `pickle`
from accessing that class, although it's not really stored in the dumped
output.